requirements.txt
streamlit
openai
psycopg2-binary
python-dotenv
pandas